requirements.txt
pandas
pyarrow
sqlalchemy
pyodbc
google-cloud-storage
google-cloud-bigquery
google-auth
tqdm